requirements.txt
numpy
tkinter
opencv-python
pillow
stegano